Move random utils functions to a dedicated file
