+++ packages/atomic/src/components/common/atomic-modal/atomic-modal.ts
@@ -11,8 +11,9 @@ import {withTailwindStyles} from '@/src/decorators/with-tailwind-styles.js';
 import {multiClassMap, tw} from '@/src/directives/multi-class-map.js';
 import {InitializeBindingsMixin} from '@/src/mixins/bindings-mixin';
 import {listenOnce} from '@/src/utils/event-utils.js';
+import {randomID} from '@/src/utils/random-utils.js';
 import {updateBreakpoints} from '@/src/utils/replace-breakpoint-utils';
-import {once, randomID} from '@/src/utils/utils.js';
+import {once} from '@/src/utils/utils.js';
 import type {AtomicFocusTrap} from '../atomic-focus-trap/atomic-focus-trap.js';
 import type {AnyBindings} from '../interface/bindings.js';
