+++ packages/atomic/src/components/common/facets/facet-value-checkbox/facet-value-checkbox.ts
@@ -2,8 +2,8 @@ import {html} from 'lit';
 import {keyed} from 'lit/directives/keyed.js';
 import {ref} from 'lit/directives/ref.js';
 import type {FunctionalComponentWithChildren} from '@/src/utils/functional-component-utils';
+import {randomID} from '../../../../utils/random-utils';
 import {createRipple} from '../../../../utils/ripple-utils';
-import {randomID} from '../../../../utils/utils';
 import {renderCheckbox as checkbox} from '../../checkbox';
 import {renderTriStateCheckbox} from '../../triStateCheckbox';
 import type {FacetValuePropsBase} from '../facet-common';
