+++ packages/atomic/src/components/insight/atomic-insight-pager/atomic-insight-pager.tsx
@@ -22,7 +22,7 @@ import {
   InitializableComponent,
   InitializeBindings,
 } from '../../../utils/initialization-utils';
-import {randomID} from '../../../utils/utils';
+import {randomID} from '../../../utils/random-utils';
 import {
   PagerNextButton,
   PagerPageButton,
