+++ packages/atomic/src/components/insight/atomic-insight-search-box/atomic-insight-search-box.tsx
@@ -13,7 +13,8 @@ import {
   InitializeBindings,
 } from '../../../utils/initialization-utils';
 import {AriaLiveRegion} from '../../../utils/stencil-accessibility-utils';
-import {isFocusingOut, randomID} from '../../../utils/utils';
+import {isFocusingOut} from '../../../utils/utils';
+import {randomID} from '../../../utils/random-utils';
 import {SearchBoxWrapper} from '../../common/search-box/stencil-search-box-wrapper';
 import {SearchTextArea} from '../../common/search-box/stencil-search-text-area';
 import {
