+++ packages/atomic/src/components/insight/smart-snippets/atomic-insight-smart-snippet/atomic-insight-smart-snippet.tsx
@@ -6,7 +6,7 @@ import {
   SmartSnippetWrapper,
 } from '@/src/components/common/smart-snippets/atomic-smart-snippet/stencil-smart-snippet-common';
 import {Hidden} from '@/src/components/common/stencil-hidden';
-import {randomID} from '@/src/utils/utils';
+import {randomID} from '@/src/utils/random-utils';
 import {
   buildSmartSnippet as buildInsightSmartSnippet,
   SmartSnippet as InsightSmartSnippet,
