+++ packages/atomic/src/components/ipx/atomic-ipx-body/atomic-ipx-body.tsx
@@ -12,7 +12,8 @@ import {
   InitializeBindings,
 } from '../../../utils/initialization-utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint-utils';
-import {once, randomID} from '../../../utils/utils';
+import {once} from '../../../utils/utils';
+import {randomID} from '../../../utils/random-utils';
 import {AnyBindings} from '../../common/interface/bindings';
 
 /**
