+++ packages/atomic/src/components/ipx/atomic-ipx-embedded/atomic-ipx-embedded.tsx
@@ -13,7 +13,8 @@ import {
   InitializeBindings,
 } from '../../../utils/initialization-utils';
 import {updateBreakpoints} from '../../../utils/replace-breakpoint-utils';
-import {once, randomID} from '../../../utils/utils';
+import {once} from '../../../utils/utils';
+import {randomID} from '../../../utils/random-utils';
 import {AnyBindings} from '../../common/interface/bindings';
 
 /**
