+++ packages/atomic/src/components/search/atomic-results-per-page/atomic-results-per-page.ts
@@ -29,7 +29,7 @@ import {errorGuard} from '@/src/decorators/error-guard.js';
 import type {InitializableComponent} from '@/src/decorators/types.js';
 import {withTailwindStyles} from '@/src/decorators/with-tailwind-styles.js';
 import {InitializeBindingsMixin} from '@/src/mixins/bindings-mixin.js';
-import {randomID} from '@/src/utils/utils.js';
+import {randomID} from '@/src/utils/random-utils.js';
 
 /**
  * The `atomic-results-per-page` component determines how many results to display per page.
