+++ packages/atomic/src/components/search/facets/color-facet-checkbox/color-facet-checkbox.ts
@@ -3,8 +3,8 @@ import {keyed} from 'lit/directives/keyed.js';
 import {ref} from 'lit/directives/ref.js';
 import type {FacetValuePropsBase} from '@/src/components/common/facets/facet-common';
 import type {FunctionalComponentWithChildren} from '@/src/utils/functional-component-utils';
+import {randomID} from '@/src/utils/random-utils';
 import {createRipple} from '@/src/utils/ripple-utils';
-import {randomID} from '@/src/utils/utils';
 
 export const renderColorFacetCheckbox: FunctionalComponentWithChildren<
   FacetValuePropsBase
