+++ packages/atomic/src/utils/utils.spec.ts
@@ -1,5 +1,5 @@
 import {beforeEach, describe, expect, it, vi} from 'vitest';
-import {aggregate, isInDocument, once, randomID} from './utils';
+import {aggregate, isInDocument, once} from './utils';
 
 describe('utils', () => {
   describe('#once', () => {
@@ -49,20 +49,6 @@ describe('utils', () => {
       const result: number = fnOnce();
 
       expect(result).toBe(123);
-    });
-  });
-
-  describe('#randomID', () => {
-    it('should place the string to prepend at the start of the id when passed', () => {
-      expect(randomID('prefix')).toMatch(/^prefix/);
-    });
-    it('should prefix id with nothing when a string to prepend is not passed', () => {
-      expect(randomID()).not.toMatch(/^undefined/);
-      expect(randomID().length).toBe(5);
-    });
-
-    it('should return two different ids when called twice', () => {
-      expect(randomID()).not.toBe(randomID());
     });
   });
