+++ packages/atomic/src/utils/utils.ts
@@ -15,26 +15,12 @@ export function once<T extends unknown[], R>(fn: (...args: T) => R) {
   };
 }
 
-export function randomID(prepend?: string, length = 5) {
-  const randomStr = Math.random()
-    .toString(36)
-    .substring(2, 2 + length);
-  if (!prepend) {
-    return randomStr;
-  }
-  return prepend + randomStr;
-}
-
-export function getRandomArbitrary(min: number, max: number) {
-  return Math.random() * (max - min) + min;
+export function parseHTML(string: string) {
+  return new window.DOMParser().parseFromString(string, 'text/html');
 }
 
 export function parseXML(string: string) {
   return new window.DOMParser().parseFromString(string, 'text/xml');
-}
-
-export function parseHTML(string: string) {
-  return new window.DOMParser().parseFromString(string, 'text/html');
 }
 
 export function isElementNode(node: Node): node is Element {
